chore: clean up api handling
